Fix imports after moving Scratch files
